Renamed the Canvas and JavaScript helpers
